DLS-10863: Upgraded play/bootstrap to the version not triggering the errors
The play & bootstrap upgraded to the version that is not triggering the breakes. A separate ticket was raised to upgrade it to the latest versions (> v9).
With the upgrade of bootstrap to above 8.4.0 the AuthRedirects.scala stopped being a member of the bootstrap.config. To address that issue the AuthRedirectsExternal.scala file has been added along with the tests for that. Links to sources attached